Remove on-failure option from request-accounts
The test was hanging in karma and I don't understand the why yet. I'll just
remove the option and comment the test since it was a big of upfront design of
my part.

When I start handling errors I'll take a closer look at it. Right now, I just
want to move forward and implement a happy path.