Add create stream amount step
The amount step is still missing the dynamic assets fetch from the user's wallet
and convert the amount from and to wei. Any amount stored in the database will
be always as close as possible from the blockchain, thus wei.

Fix redirect in the recipient step so it sends the user to the amount step on
form submit.

Add Asset record in the wallet namespace. Give data some predefined shape to
express the domain better.