Add duration step
Add a new subscription to get the amount and convert it from ether to wei.

Add two new subscriptions. First one to get the amount and convert it from ether
to wei. Last one is to get the token symbol directly.

Note for my future self. Code is starting to get boilerplaty, thus some patterns
are rising here and there. I'll take a look and possibly refactor them when
finish with the form steps.