Move defcards to test namespaces
This commit fix the error "ReferenceError: React is not defined" while running
karma-test build by moving defcards to test namespaces and explicit require
cljsjs.react.

Wherever devcards is required it need cljsjs.react to be required too. It seems
to be a limitation of shadow-cljs because figwheel doesn't have the same
issue. That said, I prefer to have React included in test namespace rather than
generating noise in production ones.