Refactor onboarding
This commit refactor onboarding to have its own view and model. Also, add a few
more tests for both namespaces.

Start using namespaced keywords instead of just prefixed keywords for re-frame
registry. The reason is most times I need to import the namespace before using
something in the registry. For example, if I want to dispatch an event named
:foo, I need to first make sure to require whatever is registering :foo in
re-frame before dispatching it. Therefore, the usage of namespaced keywords just
ensures all dependencies are load correctly for a given part of the code.

Side node, the problem above could be solved by requiring all namespaces in the
core.cljs, but overloading core is a bad idea in my opinion - I don't want to
remember that every time I add something I also have to require it in core.

Refactor mask-address to return nil if nil is given as argument.