Write stream in the local storage
This is the cheapest solution to persist a single stream per user. By a mile it
isn't the ideal but it might fill the gaps to make the UI less buggy from the
confirmation to the stream details page.

Also stop storing the ERC20 ABI in the database. The interface is the same and
this can simplify our end.