Improve devcards build configuration
This commit is one from a series of changes to improve the development workflow
using devcards.

Add devtools in the devcards build configuration so it can be run in standalone
mode.

Use deftest from devcards instead of clj.test. It will allow me to see both
tests and cards in the devcards UI.